DELIA-45638: DRM failure with 412 entitlement error on DASH playback
Reason for change: Access Token acquired from auth service only on first time DASH tune and this cached token used for further tunes,
		so this cached expiry 412 error returned from DRM server on license request.
		Now added fix to flush this cached token on 412 error scenario

Test Procedure: Tested this fix by manual hard coding with 412 error on 1st time license request, Need to test this actual fix by run STB without reboot on more days and ensure DASH tune working fine
Risks: low


DELIA-45921:[AAMP][DASH] Session Token is fetch again for unauthorized channel
Manually cherry picked from dev_sprint to resolve merge conflict https://code.rdkcentral.com/r/c/rdk/components/generic/aamp/+/44739

Reason for change: In existing code drm license request failure with error code 412 considered as expired token and again access token fetched for drm license request, also this main error code 412 returned for many reason.
Now added fix to flush cached token and requested again only on expired access token scenario(main error code 412 and status code 401)
Test Procedure: Tested this update by switching to unauthorized channel (main error code 412 and extended status code 12012) and ensured that session token not fetched again. Need to test this actual expired access token update by running stb on more than one day with dash channel and ensure that channel change will hit expired session error and recovered with new access token

Risks: low

+++ drm/AampDRMSessionManager.h
@@ -147,7 +147,7 @@ public:
 			const unsigned char * initDataPtr, uint16_t dataLength, MediaType streamType,
 			const unsigned char *contentMetadata, PrivateInstanceAAMP* aamp, AAMPEvent *e,bool isPrimarySession = false);
 
-	DrmData * getLicense(DrmData * keyChallenge, string destinationURL, long *httpError, MediaType streamType, PrivateInstanceAAMP* aamp, bool isComcastStream = false, char* licenseProxy = NULL, struct curl_slist *customHeader = NULL, DRMSystems drmSystem = eDRM_NONE);
+	DrmData * getLicense(DrmData * keyChallenge, string destinationURL, int32_t *httpError, MediaType streamType, PrivateInstanceAAMP* aamp, bool isComcastStream = false, char* licenseProxy = NULL, struct curl_slist *customHeader = NULL, DRMSystems drmSystem = eDRM_NONE);
 
 	void clearSessionData();
